backend/django-rest/requirements.txt
djangorestframework
django-cors-headers


